Win32 build - copy of code, update CMakeLists.txt

+++ src/reader.c
@@ -27,7 +27,6 @@ The entry point is reader().  */
 
 #include <limits.h>
 #include <stdio.h>
-#include <string.h>
 #include <ctype.h>
 #include "system.h"
 #include "files.h"
@@ -470,7 +469,7 @@ parse_token_decl (what_is, what_is_not)
   register int token = 0;
   register int prev;
   register char *typename = 0;
-  size_t k;
+  int k;
 
 /*   start_lineno = lineno; JF */
 
@@ -546,7 +545,7 @@ parse_start_decl ()
 void
 parse_type_decl ()
 {
-  register size_t k;
+  register int k;
   register char *name;
 /*   register int start_lineno; JF */
 
@@ -603,7 +602,7 @@ void
 parse_assoc_decl (assoc)
 int assoc;
 {
-  register size_t k;
+  register int k;
   register char *name = NULL;
 /*  register int start_lineno; JF */
   register int prev = 0;	/* JF added = 0 to keep lint happy */
@@ -1597,7 +1596,7 @@ record_rule_line ()
 int
 get_type()
 {
-  register size_t k;
+  register int k;
   register int t;
   register char *name;
